+++ open-codegen/CHANGELOG.md
@@ -7,7 +7,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 Note: This is the Changelog file of `opengen` - the Python interface of OpEn
 
-## [0.9.5] - Unreleased
+
+## [0.9.6] - 2026-03-14
+
+### Fixed
+
+- Clippy fixes (minor)
+
+
+
+## [0.9.5] - 2026-03-12
 
 ### Changed 
 
@@ -246,6 +255,8 @@ Note: This is the Changelog file of `opengen` - the Python interface of OpEn
 * Fixed `lbfgs` typo
 
 
+[0.9.5]: https://github.com/alphaville/optimization-engine/compare/opengen-0.9.5...opengen-0.9.6
+[0.9.5]: https://github.com/alphaville/optimization-engine/compare/opengen-0.9.4...opengen-0.9.5
 [0.9.4]: https://github.com/alphaville/optimization-engine/compare/opengen-0.9.3...opengen-0.9.4
 [0.9.3]: https://github.com/alphaville/optimization-engine/compare/opengen-0.9.2...opengen-0.9.3
 [0.9.2]: https://github.com/alphaville/optimization-engine/compare/opengen-0.9.1...opengen-0.9.2